Organize Tableau classes
