+++ airflow/providers/tableau/sensors/tableau_job_status.py
@@ -14,59 +14,15 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from typing import TYPE_CHECKING, Optional, Sequence
 
-from airflow.providers.tableau.hooks.tableau import (
-    TableauHook,
-    TableauJobFailedException,
-    TableauJobFinishCode,
+"""This module is deprecated. Please use :mod:`airflow.providers.tableau.sensors.tableau`."""
+
+import warnings
+
+from airflow.providers.tableau.sensors.tableau import TableauJobStatusSensor  # noqa
+
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.tableau.sensors.tableau`.",
+    DeprecationWarning,
+    stacklevel=2,
 )
-from airflow.sensors.base import BaseSensorOperator
-
-if TYPE_CHECKING:
-    from airflow.utils.context import Context
-
-
-class TableauJobStatusSensor(BaseSensorOperator):
-    """
-    Watches the status of a Tableau Server Job.
-
-    .. seealso:: https://tableau.github.io/server-client-python/docs/api-ref#jobs
-
-    :param job_id: Id of the job to watch.
-    :param site_id: The id of the site where the workbook belongs to.
-    :param tableau_conn_id: The :ref:`Tableau Connection id <howto/connection:tableau>`
-        containing the credentials to authenticate to the Tableau Server.
-    """
-
-    template_fields: Sequence[str] = ('job_id',)
-
-    def __init__(
-        self,
-        *,
-        job_id: str,
-        site_id: Optional[str] = None,
-        tableau_conn_id: str = 'tableau_default',
-        **kwargs,
-    ) -> None:
-        super().__init__(**kwargs)
-        self.tableau_conn_id = tableau_conn_id
-        self.job_id = job_id
-        self.site_id = site_id
-
-    def poke(self, context: 'Context') -> bool:
-        """
-        Pokes until the job has successfully finished.
-
-        :param context: The task context during execution.
-        :return: True if it succeeded and False if not.
-        :rtype: bool
-        """
-        with TableauHook(self.site_id, self.tableau_conn_id) as tableau_hook:
-            finish_code = tableau_hook.get_job_status(job_id=self.job_id)
-            self.log.info('Current finishCode is %s (%s)', finish_code.name, finish_code.value)
-
-            if finish_code in (TableauJobFinishCode.ERROR, TableauJobFinishCode.CANCELED):
-                raise TableauJobFailedException('The Tableau Refresh Workbook Job failed!')
-
-            return finish_code == TableauJobFinishCode.SUCCESS
